+++ modules/tests/photons_control_tests/device_finder/test_special_reference.py
@@ -376,16 +376,7 @@ describe "finding devices":
                 found.append(device)
             assert [f.serial for f in found] == [V.d3.serial]
             assert found[0].info == {
-                "cap": [
-                    "color",
-                    "not_buttons",
-                    "not_chain",
-                    "not_ir",
-                    "not_matrix",
-                    "not_multizone",
-                    "not_relays",
-                    "variable_color_temp",
-                ],
+                "cap": pytest.helpers.has_caps_list("color", "variable_color_temp"),
                 "firmware_version": "2.80",
                 "hue": V.d3.attrs.color.hue,
                 "label": "kitchen",
@@ -500,16 +491,7 @@ describe "finding devices":
                 found.append(device)
             assert [f.serial for f in found] == [V.d3.serial]
             assert found[0].info == {
-                "cap": [
-                    "color",
-                    "not_buttons",
-                    "not_chain",
-                    "not_ir",
-                    "not_matrix",
-                    "not_multizone",
-                    "not_relays",
-                    "variable_color_temp",
-                ],
+                "cap": pytest.helpers.has_caps_list("color", "variable_color_temp"),
                 "hue": V.d3.attrs.color.hue,
                 "label": "kitchen",
                 "power": "off",
@@ -530,16 +512,7 @@ describe "finding devices":
                 found.append(device)
             assert [f.serial for f in found] == [V.d3.serial]
             assert found[0].info == {
-                "cap": [
-                    "color",
-                    "not_buttons",
-                    "not_chain",
-                    "not_ir",
-                    "not_matrix",
-                    "not_multizone",
-                    "not_relays",
-                    "variable_color_temp",
-                ],
+                "cap": pytest.helpers.has_caps_list("color", "variable_color_temp"),
                 "firmware_version": "2.80",
                 "hue": V.d3.attrs.color.hue,
                 "label": "kitchen",